@@ -12,10 +12,7 @@ namespace RoundRobin.Test
         [SetUp]
         public void Setup()
         {
-            _data = new List<int>()
-            {
-                1, 2, 3, 4, 5
-            };
+            _data = [1, 2, 3, 4, 5];
         }
 
         [Test]
@@ -42,7 +39,7 @@ namespace RoundRobin.Test
                 1, 2, 2, 3, 3, 4, 4, 5, 5, 1
             };
 
-            Assert.AreEqual(mustBe, result);
+            Assert.That(mustBe,Is.EqualTo(result));
         }
         
         [Test]
@@ -66,7 +63,7 @@ namespace RoundRobin.Test
                 1, 2, 2, 3 ,4 ,5 ,1 ,2 ,2 ,3 
             };
 
-            Assert.AreEqual(mustBe, result);
+            Assert.That(mustBe, Is.EqualTo(result));
         }
 
         [Test]
@@ -92,7 +89,7 @@ namespace RoundRobin.Test
                 1, 1, 1, 2, 3, 4, 5, 1, 1, 1
             };
 
-            Assert.AreEqual(mustBe, result);
+            Assert.That(mustBe,Is.EqualTo(result));
         }
 
         
@@ -117,7 +114,7 @@ namespace RoundRobin.Test
                 5, 1, 2, 3, 4, 5,1,2,3,4
             };
 
-            Assert.AreEqual(result, mustBe);
+            Assert.That(result, Is.EqualTo(mustBe));
         }
         
         [Test]
@@ -140,7 +137,7 @@ namespace RoundRobin.Test
                 1, 2, 3, 4, 5, 1, 2, 3, 4, 5
             };
 
-            Assert.AreEqual(result, mustBe);
+            Assert.That(result,Is.EqualTo(mustBe));
         }
 
         [Test]
@@ -163,10 +160,10 @@ namespace RoundRobin.Test
 
             result.ToList().ForEach(z => { TestContext.Write($"{z},"); });
 
-            Assert.AreEqual(new List<int>()
+            Assert.That(new List<int>()
             {
                 1, 2, 3, 4, 5, 1, 2, 3, 4, 5
-            }, result);
+            }, Is.EqualTo(result));
         }
     }
 }
